SAM-14, Частично вынес ui из EditingScreenViewController в отдельный view
